epub-converter/requirements.txt
ebooklib>=0.18
rich>=13.0
Pillow>=10.0
httpx>=0.27.0
